requirements.txt
beautifulsoup4==4.12.2
futures3==1.0.0
fuzzywuzzy==0.18.0
Levenshtein==0.22.0
pandas==2.1.0
pytest==7.4.3
python-dateutil==2.8.2
python-Levenshtein==0.22.0
python-telegram-bot==21.0.1
regex==2023.10.3
requests==2.31.0
selenium==4.17.2
sqlitedict==2.1.0
Unidecode==1.3.7
